Version 2.0.6
Version 2.0.6 beta?
Updated build.bat with newest JDK.
Changed DesignVariable vspValue to be a property so it can be included in the table.
Changed JavaFXUI to include the vspValue as a column in the table and added logic to restoreState() to check if the variable name is unique without the VSPID in cases when the ID has changed.
Changed ModelCenterPlugin and ModelCenterWrapper to read the new cfd files.
Modified OpenVSP3File to add CFD file output variables in a CFDFile group under the CFD settings container.
Updated OpenVSP3Plugin to 2.0.6. Added constants and a map for the CFD file types.
Updated OpenVSP3Plugin::writeVSPScriptFile() to generate requested CFD mesh files.